Move split_object_name into system_tables module
